Fix tenant isolation, harden the query engine, and clear the code-review backlog (#12)
The localdb engine replaced Supabase's row-level security, but /api/db handed
it query descriptors authored by the browser and every identifier in one went
straight into the SQL string. A quote inside a column name defeated the org
predicate, letting an org-A user read and delete org-B's rows and reach the
users table through subqueries and embeds. Separately, org_id and role were
client-writable on profiles, so any user could move into any tenant as admin.

Identifiers are now validated against the real column list, embeds are
allowlisted and org-scoped, org_id and role are stripped from client writes,
onboarding moved to a transactional server route, and destructive queries
require a filter. 27 regression tests cover each exploit as a tenant boundary.

Also fixes the Slack daemon's arbitrary code execution and push-to-main, silent
data loss in the readings editor, ambient conditions never reaching the PDF, the
UTC "Today's Jobs" bug, missing upload limits and file ownership checks, and the
78 type errors. Suite is 48 passing, tsc clean, build green.

+++ supabase/migrations/002_erp_schema.sql
@@ -147,7 +147,7 @@ CREATE TABLE contracts (
   renewal_notice_days INT DEFAULT 60,
   inspection_frequency TEXT
                         CHECK (inspection_frequency IN ('monthly', 'quarterly', 'semi_annual', 'annual')),
-  value_cents         INT, -- annual contract value in cents
+  value_cents         BIGINT, -- annual contract value in cents
   payment_terms       TEXT DEFAULT 'net30',
   notes               TEXT,
   created_at          TIMESTAMPTZ DEFAULT NOW()
